requirements.txt
fastapi[all]
uvicorn[standard]
python_dotenv
requests
